+++ CHANGELOG.md
@@ -7,6 +7,7 @@ FlatLaf Change Log
 
 - ComboBox and Spinner: Fixed missing arrow buttons if preferred height is zero.
   Minimum width of arrow buttons is 3/4 of default width.
+- TabbedPane: Switch and close tabs on left mouse click only. (PR #595)
 
 
 ## 2.5